+++ tests/dummy/app/controllers/legacy-demo.js
@@ -400,10 +400,6 @@ export default Ember.Controller.extend({
   ],
 
   actions: {
-    logFocus(select, e) {
-      console.debug('onfocus in ', e.target.classList[0]);
-    },
-
     onSelectOpen(select) {
       select.actions.search('foo');
     },