+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Master
 
+# 1.0.0-beta.23
 - [ENHANCEMENT] Add a `defaultHighlighted` option that can be used to customize what item is highlighted by default with the component is opened.
   It can either be a value, or a function that gets called with the select and returns that value.
 - [ENHANCEMENT] Add `onblur` event for symmetry with `onfocus`, and also clarify that both are fired for any element of